+++ docs/CHANGELOG.md
@@ -10,6 +10,7 @@ Version 1.9.2 (not released yet)
 - Add `af_killsound1.wav` from Alpine Faction
 - Add `hit_sounds` command
 - Add `weapon_screen_shake` command
+- Add `full_bright_entities` command
 - Add `remote_server_flags` command
 
 Version 1.9.1 (released 2025-07-05)